+++ src/topfd/feature_detect/env_set/env_set.cpp
@@ -49,7 +49,7 @@ toppic::Xic toppic::EnvSet::init_median_xic(){
     theoretical_peak_sum = theoretical_peak_sum + (theo_envelope_inte[refer_idx] * ratio);
     if (refer_idx - 1 >= 0)
       theoretical_peak_sum = theoretical_peak_sum + (theo_envelope_inte[refer_idx-1] * ratio);
-    if (refer_idx + 1 < theo_envelope_inte.size())
+    if (refer_idx + 1 < static_cast<int>(theo_envelope_inte.size()))
       theoretical_peak_sum = theoretical_peak_sum + (theo_envelope_inte[refer_idx+1] * ratio);
     inte_list.push_back(theoretical_peak_sum);
   }
@@ -120,7 +120,7 @@ void toppic::EnvSet::remove_non_consecutive_peaks(int i, int max_miss_peak){
   // search forward
   idx = seed_env_.getSpecId() - start_spec_id_;
   miss_num = 0;
-  while (idx < exp_env_list_.size()) {
+  while (idx < static_cast<int>(exp_env_list_.size())) {
     std::vector<ExpPeak> peaks = exp_env_list_[idx].getExpEnvList();
     if (peaks[i].isEmpty())
       miss_num = miss_num + 1;
@@ -130,7 +130,7 @@ void toppic::EnvSet::remove_non_consecutive_peaks(int i, int max_miss_peak){
       break;
     idx = idx + 1;
   }
-  while (idx < exp_env_list_.size()) {
+  while (idx < static_cast<int>(exp_env_list_.size())) {
     std::vector<ExpPeak> peaks = exp_env_list_[idx].getExpEnvList();
     peaks[i] = ExpPeak();
     exp_env_list_[idx].setExpEnvList(peaks);
@@ -158,12 +158,12 @@ void toppic::EnvSet::get_weight_mz_error(double* weight_sum, double* error_sum){
   std::vector<double> inte_list = xic_.getInteList();
   *weight_sum = 0;
   *error_sum = 0;
-  for (int spec_idx = 0; spec_idx < exp_env_list_.size(); spec_idx++) {
+  for (size_t spec_idx = 0; spec_idx < exp_env_list_.size(); spec_idx++) {
     ExpEnvelope expEnvelope = exp_env_list_[spec_idx];
     if (expEnvelope.isEmpty())
       continue;
     std::vector<ExpPeak> exp_peak_list = expEnvelope.getExpEnvList();
-    for (int peak_idx = 0; peak_idx < seed_peak_list.size(); peak_idx++) {
+    for (size_t peak_idx = 0; peak_idx < seed_peak_list.size(); peak_idx++) {
       ExpPeak peak = exp_peak_list[peak_idx];
       if (!peak.isEmpty()) {
         double cur_inte = seed_peak_list[peak_idx].getInte() * inte_list[spec_idx];
@@ -199,7 +199,7 @@ double get_left_max(int pos, std::vector<double> &y) {
 
 double get_right_max(int pos, std::vector<double> &y){
   double max_val = -100000000;
-  for (int i = pos + 1; i < y.size(); i++) {
+  for (size_t i = pos + 1; i < y.size(); i++) {
     if (y[i] > max_val)
       max_val = y[i];
   }
@@ -213,7 +213,7 @@ void toppic::EnvSet::shortlistExpEnvs() {
   std::vector<double> shortlisted_inte_list;
   std::vector<double> shortlisted_smoothed_inte_list;
   std::vector<ExpEnvelope> tmp;
-  for (int i = 0; i < exp_env_list_.size(); i++) {
+  for (size_t i = 0; i < exp_env_list_.size(); i++) {
     if (exp_env_list_[i].getSpecId() >= start_spec_id_ and exp_env_list_[i].getSpecId() <= end_spec_id_) {
       tmp.push_back(exp_env_list_[i]);
       shortlisted_inte_list.push_back(inte_list[i]);
@@ -307,7 +307,7 @@ double toppic::EnvSet::comp_intensity(double snr, double noise_inte){
   std::vector<std::vector<double>> map = this->get_map(snr, noise_inte);
   std::vector<double> aggregate_inte (map[0].size(), 0.0);
   for (auto &sp_map : map)
-    for (int peakIdx = 0; peakIdx < aggregate_inte.size(); peakIdx++)
+    for (size_t peakIdx = 0; peakIdx < aggregate_inte.size(); peakIdx++)
         aggregate_inte[peakIdx] = aggregate_inte[peakIdx] + sp_map[peakIdx];
   double abundance = std::accumulate(aggregate_inte.begin(), aggregate_inte.end(), 0.0);
   return abundance;
@@ -349,14 +349,14 @@ void toppic::EnvSet::remove_matrix_peaks(PeakMatrix &peak_matrix){
 
 void toppic::EnvSet::remove_peak_data(PeakMatrix &peakMatrix) {
   std::vector<std::vector<double>> map = get_map(3.0, peakMatrix.get_min_inte());
-  for (int env_id = 0; env_id < exp_env_list_.size(); env_id++) {
+  for (size_t env_id = 0; env_id < exp_env_list_.size(); env_id++) {
     ExpEnvelope exp_env = exp_env_list_[env_id];
     int spec_id = exp_env.getSpecId();
     if (spec_id < 0 or spec_id >= peakMatrix.get_spec_num())
       continue;
     std::vector<ExpPeak> exp_data = exp_env.getExpEnvList();
     std::vector<double> theo_data = map[env_id];
-    for (int elem_idx = 0; elem_idx < exp_data.size(); elem_idx++) {
+    for (size_t elem_idx = 0; elem_idx < exp_data.size(); elem_idx++) {
       ExpPeak exp_peak = exp_data[elem_idx];
       if (exp_peak.isEmpty())
         continue;
